solution of the issue 24

+++ src/perturbopy/test_utils/run_test/run_utils.py
@@ -492,19 +492,38 @@ def ph_collection(prefix, nq_num):
     for file in os.listdir('.'):
         if file.startswith(f'{prefix}.dyn'):
             shutil.copy(file, 'save')
-
+    print('before the cicle')
     for nq in range(1, nq_num + 1):
         
         # copy dvscf files
         if nq > 1:
-            shutil.copy(f'{dir}/{prefix}.q_{nq}/{prefix}.dvscf1', f'save/{prefix}.dvscf_q{nq}')
+
+            # parallel version
+            if os.path.exists(f'{dir}/{prefix}.q_{nq}/{prefix}.dvscf1'):
+                shutil.copy(f'{dir}/{prefix}.q_{nq}/{prefix}.dvscf1', f'save/{prefix}.dvscf_q{nq}')
+            # serial version
+            elif os.path.exists(f'{dir}/{prefix}.q_{nq}/{prefix}.dvscf'):
+                shutil.copy(f'{dir}/{prefix}.q_{nq}/{prefix}.dvscf', f'save/{prefix}.dvscf_q{nq}')
+            else:
+                raise FileNotFoundError(f"{dir}/{prefix}.q_{nq}/{prefix}.dvscf and {dir}/{prefix}.q_{nq}/{prefix}.dvscf1 don't exist")
+
             for file in os.listdir(f'{dir}/{prefix}.q_{nq}'):
                 if file.endswith('wfc'):
                     os.remove(f'{dir}/{prefix}.q_{nq}/{file}')
+
         else:
-            shutil.copy(f'{dir}/{prefix}.dvscf1', f'save/{prefix}.dvscf_q{nq}')
+
+            # parallel version
+            if os.path.exists(f'{dir}/{prefix}.dvscf1'):
+                shutil.copy(f'{dir}/{prefix}.dvscf1', f'save/{prefix}.dvscf_q{nq}')
+            # serial version
+            elif os.path.exists(f'{dir}/{prefix}.dvscf'):
+                shutil.copy(f'{dir}/{prefix}.dvscf', f'save/{prefix}.dvscf_q{nq}')
+            else:
+                raise FileNotFoundError(f"{dir}/{prefix}.dvscf and {dir}/{prefix}.dvscf1 don't exist")
+
             for file in os.listdir(dir):
                 if 'wfc' in file:
                     os.remove(f'{dir}/{file}')
-                    
+             
     shutil.copy(f'save/{prefix}.dyn0', f'save/{prefix}.dyn0.xml')
